Move generated metadata files into their own directory
