+++ metaspace/webapp/src/lib/metadataRegistry.ts
@@ -1,11 +1,11 @@
-import metadataMapping from './metadataMapping';
+import metadataMapping from './metadataSchemas/metadataMapping';
 
 export const metadataTypes = Object.keys(metadataMapping);
 
 export const defaultMetadataType = metadataTypes.includes("Imaging MS") ? "Imaging MS" : metadataTypes[0];
 
-export const metadataSchemas = {};
+export const metadataSchemas: Record<string, any> = {};
 for (const mdType of metadataTypes) {
   const mdFilename = metadataMapping[mdType];
-  metadataSchemas[mdType] = require(`../assets/${mdFilename}`);
+  metadataSchemas[mdType] = require(`./metadataSchemas/${mdFilename}`);
 }
